automatically manage ZLS versions
These changes make ZLS an "invisible" component of the extension after the initial setup.

This also removes the `zig.zls.checkForUpdate` command. I am not sure if it even needs to be brought back. I've made sure that no error is reported if no internet connection is available. The last installed ZLS version should be reused.

closes #136 because `zig.zls.startRestart` will just enable ZLS instead of complaining