Set cwd when running tests + no-llvm option
When running tests the current working directory of the process is set to the
workspace folder containing the test, which is the project folder. This allows
tests to open files relative to the project.

For example: `std.fs.cwd().openFile("./tests/input", .{})`
This mirrors running `zig build test` or `zig test` on the command line within the project folder.

A new configurable option "no-llvm" can be enabled to enable fast compilations
by switching the code generation from llvm to the x86 backend. Note this is
experimental and is not available on all machine targets.