use `URL` instead of `vscode.Uri` to avoid unnecessary percent encoding
The `toString` function on `vscode.Uri` will aggressively apply percent
encoding even if unnecessary for some character+component cases.

This affects the `+` in the filename of nightly versions and the `=` in
the query parameter. While it is valid to percent encode even if
unnecessary, this can still lead to confusion about whether the url
is valid if the reader is unfamiliar with the rules around
percent encoding.

There is no downside to just using `URL` instead of `vscode.Uri` so
that we get more readable urls. This can be especially helpful if the
encoded url happens to be shown to the extension user through the
error message of an exception as an example.